Moved FluentMetric to metric.py
I need to add a significant new class in the next commit, but it's
impossible to do so cleanly without proper python project structure.
This change moves FluentMetric classes to metric.py and then simply
imports them in __init__.py, which effectively exports them. There is no
breaking change in this commit. It's just easier to add more files now.
